Major reorganization to files and bug fixes.
Reorganized the files.
Modified the button more after noticing an issue with parse's location.
Added new files for better management. Questioning the need for folders
though feel they may come in handy at some point... either that or I
will just delete them. Not sure, I like the idea of consistent
organization, but a directory with one item seems pointless.
Laid the groundwork for parse(). Super barebones.
Fixed issues with Lex involving improper Keywords, and removed
subtraction support as it was not on the grammar (thought it was a
mistake, though assuming now it was not. It is still supported, just
commented out!).