Fixed issue with lex/fixed parse.
Finished Project 1 (for the most part. Not happy about parse(), but
would rather go back when I have time then submit a broken version(more
broken)).
Fixed Bool issues in parse, though there are easily going to be more.
Not going to lie, this was quite a bit of a hassle, wish I had started
when most of my classes were in more of a lull, but it never works out
when hindsight is 20/20.