Attempt 3 at parse. First "real" attempt...
Commented out old parse code.
Basic steps in the right direction, I think...

(Wish I had a better handle in terms of going about coding this properly
as I keep feeling like I am missing some key idea. Wish I had more time
in life in general as it would probably help. Hope this is the last
restart, though, this is humorous to me, as if this attempt fails, it
will be like the last few, this time though, I am committing as it makes
me feel like I just cannot toss all of the work when it seems to break
after thinking it should work. First goal is getting the program "{}$"
to be a success. I hope to have this done by Friday/Saturday, but
actually get it done this time. The main reason this is humorous is that
no one will see this unless I stick with this version, and at that point
none of this writing will matter. Oh life.)