Clean up and progress on Symbol Table.
Started really laying out how the Symbol Table will work.
Int Expression more or less done.
Added basis error outputting for ST.
Fixed issue noticed in AST (said it was building a CST when it was
building an AST).
Added a new file which is the old version of a function formally used in
ST building which is not, but just in case thought I should leave it.