Ready for project 2 submission (I think).
Removed 2 buttons.
Modified 2 to contain some of Alan's tests.
Modified the overall compiler to not display the CST or AST if there is
a Symbol Table error.
Changed testdata.txt to now be a run of Alan's massive program (aka the
built in test program.
Changed testprograms.txt to include Alan's current test programs.
Removed js\ast\old.txt as it was officially no longer needed.